First step into replacing the path-resolver with a src-resolver.
The CJS module spec is agnostic about how modules are stored
and so should we.

Real-world use case: native node module src code is
available through a C++ binding, not the file system.
These changes will allow us to subclass/monkey-patch
srcParser to graph native node modules should we
want to (or deal with modules stored in a db).