Properly handle dir modules. A module `foo` which points to `foo/index` is now handled by creating two modules:  `foo` and `foo/index`. The content of `foo` is generated so that it just points to `foo/index` using node specific syntax (which is okay as dir modules are also node-specific):
    // Generated content in module `foo`
    module.exports = require('foo/index');

This has the main benefit of keeping the context correct when
dealing with relative module identifiers. It's also rather
consistent with the semantics of dir modules.

The main issue with this solution however, is that it artificially
adds an extra level of indirection. Thankfully this is something
we can handle at a later stage by comparing either the raw attribute
of modules, or their fullpath and acting accordingly.