Allow optional caching of parse results
Right now each javascript file is read and parsed to figure out the
arguments of require(..) in those files.  This happens on every build
even though most files do not change.  It would be far more efficient to
cache those results to use on recompiles, as the parsing step is slow.